+++ lib/components/segmented-control/segmented-control.scss
@@ -12,13 +12,18 @@
         align-items: center;
         cursor: pointer;
 
-        &:hover {
+        &:hover:not(.disabled) {
             background: var(--core-color-opacity-black-100);
         }
-        &:active {
+        &:active:not(.disabled) {
             background: var(--core-color-opacity-black-200);
         }
-        &.selected {
+        &.disabled {
+            opacity: 0.24;
+            cursor: not-allowed;
+            pointer-events: none;
+        }
+        &.selected:not(.disabled) {
             background: var(--core-color-solid-slate-50);
         }
         .icon {
